requirements.txt
attrs==22.1.0
aws-cdk-lib==2.50.0
aws-cdk.aws-lambda-python-alpha==2.50.0a0
boto3==1.21.11
botocore==1.24.11
cattrs==22.2.0
constructs==10.1.162
datadog-cdk-constructs-v2==1.0.0
exceptiongroup==1.0.4
jmespath==0.10.0
jsii==1.71.0
publication==0.0.3
python-dateutil==2.8.2
s3transfer==0.5.2
six==1.16.0
typeguard==2.13.3
typing_extensions==4.4.0
urllib3==1.26.8